Fix enum typing bug from commit cedbd78451
A glong/gint mismatch was wreaking havoc on various enums.
It's clear why this was broken; g_value_info_get_value says
that it returns a glong.

+++ libseed/seed-importer.c
@@ -122,7 +122,7 @@ seed_gi_importer_handle_enum (JSContextRef ctx,
   JSObjectRef enum_class;
   guint num_vals, i, j;
   gsize name_len;
-  glong value;
+  gint value; // TODO: investigate what's up with the glong/gint mystery here
   gchar *name;
   GIValueInfo *val;
